add a bunch of documentation and some preliminary stuff around variable types
