+++ Sources/ApolloCodegenLib/ASTArgument.swift
@@ -1,7 +1,8 @@
 import Foundation
 
+/// An argument which can be 
 class ASTArgument: Codable {
   let name: String
   let value: JSONContainer
-  let type: String
+  let type: ASTVariableType
 }
